Create AuthenticationConnection.py
Code to authenticate and connect to the Twitter API using the tweepy library.


Replace YOUR_CONSUMER_KEY, YOUR_CONSUMER_SECRET, YOUR_ACCESS_TOKEN, and YOUR_ACCESS_TOKEN_SECRET with the corresponding API credentials that you obtained from Twitter. The code uses the OAuthHandler class to authenticate with the Twitter API and the API class to create an API object for accessing the API. Finally, the code verifies the API credentials by calling the verify_credentials() method.